# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pelayanan Publik API",
        "description": "API Backend untuk Aplikasi Pelayanan Publik",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/berita": {
            "get": {
                "tags": [
                    "Berita"
                ],
                "summary": "Get all berita",
                "operationId": "106b2395ee105cc16df2ca30510011b8",
                "parameters": [
                    {
                        "name": "kategori",
                        "in": "query",
                        "description": "Filter by kategori ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by title or content",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of berita"
                    }
                }
            }
        },
        "/berita/{id}": {
            "get": {
                "tags": [
                    "Berita"
                ],
                "summary": "Get single berita",
                "operationId": "5cc8f589f24b393b48692be6850c829e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berita details"
                    }
                }
            }
        },
        "/admin/berita": {
            "post": {
                "tags": [
                    "Berita"
                ],
                "summary": "Create new berita",
                "operationId": "8875d632720556b90651f94fd5bfb714",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "judul": {
                                        "type": "string"
                                    },
                                    "isi": {
                                        "type": "string"
                                    },
                                    "gambar": {
                                        "type": "file"
                                    },
                                    "id_kategori": {
                                        "type": "integer"
                                    },
                                    "is_published": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berita created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/berita/{id}": {
            "put": {
                "tags": [
                    "Berita"
                ],
                "summary": "Update berita",
                "operationId": "6adbc479240cbad1ea0d75c13ae4a37a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berita updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Berita"
                ],
                "summary": "Delete berita",
                "operationId": "f5d55c78188f6b095a202e237414d49e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berita deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}